checkpoint before checking out refactor
