+++ TrafficLightsPlus.xcodeproj/xcuserdata/ncctechnology.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>SchemeUserState</key>
+	<dict>
+		<key>TrafficLightsPlus.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>0</integer>
+		</dict>
+		<key>TrafficLightsPlusTests.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>1</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>F45C7BF7300A01A2008CD241</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>F45C7C07300A01A3008CD241</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>F45C7C11300A01A3008CD241</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
+</dict>
+</plist>
