+++ TrafficLightsPlus/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,0 +1,74 @@
+{
+  "images" : [
+    {
+      "size" : "1024x1024",
+      "scale" : "1x",
+      "filename" : "icon_1024.png",
+      "idiom" : "ios-marketing"
+    },
+    {
+      "idiom" : "mac",
+      "scale" : "1x",
+      "size" : "16x16",
+      "filename" : "icon_16.png"
+    },
+    {
+      "idiom" : "mac",
+      "size" : "16x16",
+      "scale" : "2x",
+      "filename" : "icon_32.png"
+    },
+    {
+      "idiom" : "mac",
+      "filename" : "icon_32.png",
+      "size" : "32x32",
+      "scale" : "1x"
+    },
+    {
+      "idiom" : "mac",
+      "filename" : "icon_64.png",
+      "scale" : "2x",
+      "size" : "32x32"
+    },
+    {
+      "idiom" : "mac",
+      "scale" : "1x",
+      "filename" : "icon_128.png",
+      "size" : "128x128"
+    },
+    {
+      "filename" : "icon_256.png",
+      "scale" : "2x",
+      "size" : "128x128",
+      "idiom" : "mac"
+    },
+    {
+      "idiom" : "mac",
+      "scale" : "1x",
+      "size" : "256x256",
+      "filename" : "icon_256.png"
+    },
+    {
+      "filename" : "icon_512.png",
+      "idiom" : "mac",
+      "scale" : "2x",
+      "size" : "256x256"
+    },
+    {
+      "size" : "512x512",
+      "idiom" : "mac",
+      "filename" : "icon_512.png",
+      "scale" : "1x"
+    },
+    {
+      "size" : "512x512",
+      "scale" : "2x",
+      "filename" : "icon_1024.png",
+      "idiom" : "mac"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
