+++ TrafficLightsPlus/en.lproj/Localizable.strings
@@ -1,0 +1,48 @@
+"app.menu.settings" = "Settings…";
+"app.menu.enabled" = "Enabled";
+"app.menu.quit" = "Quit";
+"app.settingsWindow.title" = "Traffic Lights Plus Settings";
+
+"behavior.closeWindow" = "Close Window";
+"behavior.quitApplication" = "Quit Current App";
+"behavior.minimizeWindow" = "Minimize Window";
+"behavior.zoomWindow" = "Zoom Window";
+"behavior.hideApplication" = "Hide Current App";
+"behavior.doNothing" = "Do Nothing";
+
+"style.macOS" = "macOS Circles";
+"style.edgeSquares" = "Left Edge Squares";
+"reveal.group" = "Group";
+"reveal.nearest" = "Single (Recommended)";
+
+"settings.hiddenTrafficLights" = "Hidden Traffic Lights (Recommended)";
+"settings.unableToAddApplication" = "Unable to Add App";
+"common.ok" = "OK";
+"settings.tagline" = "Larger, easier-to-use window controls";
+"settings.enabled" = "Enabled";
+"settings.livePreview" = "Live Preview";
+"settings.appearance" = "Appearance";
+"settings.buttonSize" = "Button Size";
+"settings.resetDefault" = "Reset to Default";
+"settings.buttonSpacing" = "Button Spacing";
+"settings.restoreSystemSpacing" = "Restore System Spacing";
+"settings.revealMethod" = "Reveal Method";
+"settings.showInFullScreen" = "Enable Traffic Lights in Full-Screen Windows";
+"settings.buttonActions" = "Button Actions";
+"settings.redButton" = "Red Button";
+"settings.yellowButton" = "Yellow Button";
+"settings.greenButton" = "Green Button";
+"settings.quitOnCloseApplications" = "Apps to Quit When Closing";
+"settings.addApplication" = "Add App";
+"settings.addQuitOnCloseApplication" = "Add an app to quit when closing";
+"settings.noApplications" = "No apps added";
+"settings.removeFromList" = "Remove from list";
+"settings.removeApplication" = "Remove %@";
+"settings.systemSpacing" = "System";
+"settings.accessibilityGranted" = "Accessibility access is enabled";
+"settings.accessibilityRequired" = "Accessibility access is required to control windows";
+"settings.openAccessibilitySettings" = "Open Accessibility Settings";
+"settings.accessibilityInstructions" = "If this app is not listed, click “+” and select Traffic Lights Plus.app.";
+"settings.chooseQuitOnCloseApplication" = "Choose an App to Quit When Closing";
+"settings.add" = "Add";
+"settings.missingBundleIdentifier" = "The selected app does not have a recognizable Bundle ID.";
