+++ TrafficLightsPlus/zh-Hans.lproj/Localizable.strings
@@ -1,0 +1,48 @@
+"app.menu.settings" = "设置…";
+"app.menu.enabled" = "启用";
+"app.menu.quit" = "退出";
+"app.settingsWindow.title" = "Traffic Lights Plus 设置";
+
+"behavior.closeWindow" = "关闭窗口";
+"behavior.quitApplication" = "退出当前应用";
+"behavior.minimizeWindow" = "最小化窗口";
+"behavior.zoomWindow" = "缩放窗口";
+"behavior.hideApplication" = "隐藏当前应用";
+"behavior.doNothing" = "无操作";
+
+"style.macOS" = "macOS 圆形";
+"style.edgeSquares" = "左侧贴边方块";
+"reveal.group" = "整组";
+"reveal.nearest" = "单个（推荐）";
+
+"settings.hiddenTrafficLights" = "隐藏式红绿灯（推荐）";
+"settings.unableToAddApplication" = "无法添加应用";
+"common.ok" = "好";
+"settings.tagline" = "更大、更顺手的窗口控制按钮";
+"settings.enabled" = "启用";
+"settings.livePreview" = "实时预览";
+"settings.appearance" = "外观";
+"settings.buttonSize" = "按钮大小";
+"settings.resetDefault" = "恢复默认";
+"settings.buttonSpacing" = "按钮间距";
+"settings.restoreSystemSpacing" = "恢复系统间距";
+"settings.revealMethod" = "放大方式";
+"settings.showInFullScreen" = "处理全屏窗口的交通灯按钮";
+"settings.buttonActions" = "按钮功能";
+"settings.redButton" = "红色按钮";
+"settings.yellowButton" = "黄色按钮";
+"settings.greenButton" = "绿色按钮";
+"settings.quitOnCloseApplications" = "关闭时退出的应用";
+"settings.addApplication" = "添加应用";
+"settings.addQuitOnCloseApplication" = "添加关闭时退出的应用";
+"settings.noApplications" = "尚未添加应用";
+"settings.removeFromList" = "从名单中移除";
+"settings.removeApplication" = "移除 %@";
+"settings.systemSpacing" = "系统";
+"settings.accessibilityGranted" = "辅助功能权限已开启";
+"settings.accessibilityRequired" = "需要辅助功能权限才能控制窗口";
+"settings.openAccessibilitySettings" = "打开辅助功能设置";
+"settings.accessibilityInstructions" = "若列表中没有本应用，请点击“+”并选择 Traffic Lights Plus.app。";
+"settings.chooseQuitOnCloseApplication" = "选择关闭时退出的应用";
+"settings.add" = "添加";
+"settings.missingBundleIdentifier" = "所选应用没有可识别的 Bundle ID。";
