+++ TrafficLightsPlusTests/TrafficLightsPlusTests.swift
@@ -1,0 +1,13 @@
+import Testing
+
+@testable import TrafficLightsPlus
+
+struct TrafficLightsPlusTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+        // Swift Testing Documentation
+        // https://developer.apple.com/documentation/testing
+    }
+
+}
